Feature/refactor norms (#267)
* simplify template-based logic

* simplify feature checks during indexation

* optimize features

* avoid using set for norm columns deduplication

* correctly track payload

* introduce index features - part 0

* introduce index features - part 1

* wip

* wip

* wip

* cleanup

* introduce index features - part 2

* cleanup

* wip

* wip

* adjust benchmarks

* wip

* add pluggable field features

* wip

* wip

* wip

* fix tests

* fix tests

* wip

* adjust benchmarks

* wip

* allow pluggability of field features

* fix tests

* cleanup

* cleanup

* fix tests

* extend benchmark tests

* fix tests

* fix tests

* add fixed length norms

* address msvc build issues

* cleanup

* fix benchmarks

* cleanup

* add test

* wip

* add more tests, ensure finally::~finally() is noexcept

* add tests for misc utils

* allow using cached_func in constexpr context

* add tests for tfidf norms

* make bm25 aware of fixed length norms

* cleanup

* fix msvc build

* fix debug tests

* minor cleanup

* update CHANGELOG and README

* adjust README

* adjust README

* automatically expand tilda in VPACK_ROOT

* address review comments

* address review comments

* address review comments

+++ core/analysis/text_token_stemming_stream.cpp
@@ -301,7 +301,7 @@ bool text_token_stemming_stream::reset(const irs::string_ref& data) {
   // find the token stem
   // ...........................................................................
   if (stemmer_) {
-    if (term_buf_ref.size() > std::numeric_limits<int>::max()) {
+    if (term_buf_ref.size() > static_cast<uint32_t>(std::numeric_limits<int>::max())) {
       IR_FRMT_WARN(
         "Token size greater than the supported maximum size '%d', truncating token: %s",
         std::numeric_limits<int>::max(), data.c_str());
